chatbot_service/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-groq
langchain-community
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
python-decouple